Include base tiles of Spanish public administrations on a leaflet map

Include tiles of public Spanish organisms to a leaflet::leaflet() map.

Usage

addProviderEspTiles(
  map,
  provider,
  layerId = NULL,
  group = NULL,
  options = providerEspTileOptions()
)

providerEspTileOptions(...)

Código

https://dieghernan.github.io/leaflet-providersESP/ leaflet plugin, v1.2.0.

Argumentos

map

A map widget created from leaflet::leaflet().

provider

Name of the provider, see leaflet.providersESP.df for values available.

layerId

the layer id

group

The name of the group the newly created layers should belong to Human-friendly group names are permitted–they need not be short, identifier-style names. Any number of layers and even different types of layers (e.g. markers and polygons) can share the same group name. See leaflet::addTiles().

options

a list of extra options for tile layers, popups, paths (circles, rectangles, polygons, ...), or other map elements

...

Arguments passed on to leaflet::providerTileOptions

errorTileUrl

the tile layer options; see https://leafletjs.com/reference-1.3.4.html#tilelayer

noWrap

the tile layer options; see https://leafletjs.com/reference-1.3.4.html#tilelayer

opacity

the tile layer options; see https://leafletjs.com/reference-1.3.4.html#tilelayer

zIndex

the tile layer options; see https://leafletjs.com/reference-1.3.4.html#tilelayer

updateWhenIdle

the tile layer options; see https://leafletjs.com/reference-1.3.4.html#tilelayer

detectRetina

the tile layer options; see https://leafletjs.com/reference-1.3.4.html#tilelayer

Valor

A map object generated with leaflet::leaflet().

Detalles

providerEspTileOptions() is a wrapper of leaflet::providerTileOptions().

Véa también

Ejemplos

library(leaflet)
PuertadelSol <-
  leaflet() %>%
  setView(
    lat = 40.4166,
    lng = -3.7038400,
    zoom = 18
  ) %>%
  addProviderEspTiles(provider = "IGNBase.Gris") %>%
  addProviderEspTiles(provider = "RedTransporte.Carreteras")

PuertadelSol