Check access to SIANE data resources

Check whether R has access to resources at https://github.com/rOpenSpain/mapSpain/tree/sianedata.

Usage

esp_check_access()

Value

Logical scalar, TRUE if accessible and FALSE otherwise.

See also

Examples

esp_check_access()
#> [1] TRUE

Site built with pkgdown 2.2.0.

Template rostemplate by dieghernan, based on Universal Theme for Hugo.