Clear your mapSpain cache dir

Use this function with caution. This function would clear your cached data and configuration, specifically:

Deletes the mapSpain config directory (rappdirs::user_config_dir("mapSpain", "R")).

Deletes the cache_dir directory.

Deletes the values on stored on Sys.getenv("MAPSPAIN_CACHE_DIR") and options(mapSpain_cache_dir).

Usage

esp_clear_cache(config = FALSE, cached_data = TRUE, verbose = FALSE)

Arguments

config

if TRUE, will delete the configuration folder of mapSpain.

cached_data

If this is set to TRUE, it will delete your cache_dir and all its content.

verbose

Logical, displays information. Useful for debugging, default is FALSE.

Value

Invisible. This function is called for its side effects.

Details

This is an overkill function that is intended to reset your status as it you would never have installed and/or used mapSpain.

See also

Examples

# Don't run this! It would modify your current state
# \dontrun{
esp_clear_cache(verbose = TRUE)
#> mapSpain cached data deleted: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpC3Iv0C/mapSpain
# }

Sys.getenv("MAPSPAIN_CACHE_DIR")
#> [1] ""

Site built with pkgdown 2.0.6.

Template rostemplate by dieghernan, based on Bootstrapious.