Spanish Code Translation Data Frame

A data frame used internally for translating codes and names of the different subdivisions of Spain. The data frame provides the hierarchy of the subdivisions including NUTS1 level, Autonomous Communities (equivalent to NUTS2), Provinces and NUTS3 level. See Note.

Format

data frame with codes as columns

nuts*.code: NUTS code of each subdivision.

nuts*.code: NUTS name of each subdivision.

codauto: INE code of each autonomous community.

iso2.*.code: ISO2 code of each autonomous community and province.

ine.*.name: INE name of each autonomous community and province.

iso2.*.name.*: ISO2 name of each autonomous community and province. Several languages available.

cldr.*.name.*: CLDR name of each autonomous community and province. Several languages available.

ccaa.short.*: Short (common) name of each autonomous community. Several languages available.

cpro: INE code of each province.

prov.shortname.*: Short (common) name of each province. Several languages available.

Source

INE: Instituto Nacional de Estadistica: https://www.ine.es/

Eurostat (NUTS): https://ec.europa.eu/eurostat/web/nuts/background

ISO: https://www.iso.org/obp/ui/#iso:code:3166:ES

CLDR: https://unicode-org.github.io/cldr-staging/charts/38/index.html

Note

Languages available are:

en: English

es: Spanish

ca: Catalan

ga: Galician

eu: Basque

Although NUTS2 matches the first subdivision level of Spain (CCAA - Autonomous Communities), it should be noted that NUTS3 does not match the second subdivision level of Spain (Provinces). NUTS3 provides a dedicated code for major islands whereas the Provinces doesn't.

Ceuta and Melilla has an specific status (Autonomous Cities) but are considered as communities with a single province (as Madrid, Asturias or Murcia) on this dataset.

Examples

data(esp_codelist)