NUTS 2024 for Spain sf object

This dataset represents the spanish regions for levels 0, 1, 2 and 3 of the Nomenclature of Territorial Units for Statistics (NUTS) for 2024.

Format

A sf object with MULTIPOLYGON geometries, resolution: 1:1 million and EPSG:4258. with 86 rows and 10 variables:

NUTS_ID

NUTS identifier.

LEVL_CODE

NUTS level code (0,1,2,3).

CNTR_CODE

Eurostat Country code.

NAME_LATN

NUTS name on Latin characters.

NUTS_NAME

NUTS name on local alphabet.

MOUNT_TYPE

Mount Type, see Details.

URBN_TYPE

Urban Type, see Details.

COAST_TYPE

Coast Type, see Details.

geo

Same as NUTS_ID, provided for compatibility with eurostat.

geometry

geometry field.

Source

NUTS_RG_01M_2024_4326.gpkg file.

Details

MOUNT_TYPE: Mountain typology:

1: More than 50 % of the surface is covered by topographic mountain areas.

2: More than 50 % of the regional population lives in topographic mountain areas.

3: More than 50 % of the surface is covered by topographic mountain areas and where more than 50 % of the regional population lives in these mountain areas.

4: Non-mountain region / other regions.

0: No classification provided.

URBN_TYPE: Urban-rural typology:

1: Predominantly urban region.

2: Intermediate region.

3: Predominantly rural region.

0: No classification provided.

COAST_TYPE: Coastal typology:

1: Coastal (on coast).

2: Coastal (less than 50% of population living within 50 km. of the coastline).

3: Non-coastal region.

0: No classification provided.

See also

Examples

data("esp_nuts_2024")
head(esp_nuts_2024)
#> Simple feature collection with 6 features and 9 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -18.15996 ymin: 27.63846 xmax: 4.320228 ymax: 43.78924
#> Geodetic CRS:  ETRS89
#> # A tibble: 6 × 10
#>   NUTS_ID LEVL_CODE CNTR_CODE NAME_LATN           NUTS_NAME MOUNT_TYPE URBN_TYPE
#>   <chr>       <int> <chr>     <chr>               <chr>          <int>     <int>
#> 1 ES              0 ES        España              España            NA        NA
#> 2 ES1             1 ES        Noroeste            Noroeste          NA        NA
#> 3 ES2             1 ES        Noreste             Noreste           NA        NA
#> 4 ES3             1 ES        Comunidad de Madrid Comunida…         NA        NA
#> 5 ES4             1 ES        Centro (ES)         Centro (…         NA        NA
#> 6 ES5             1 ES        Este                Este              NA        NA
#> # ℹ 3 more variables: COAST_TYPE <int>, geo <chr>, geometry <MULTIPOLYGON [°]>

Site built with pkgdown 2.2.0.

Template rostemplate by dieghernan, based on bootstraptemple.com.