Public WMS and WMTS of Spain

This data frame is not longer in use by mapSpain. See esp_tiles_providers instead.

A data frame containing information of different public WMS and WMTS providers of Spain

Format

A data frame object with a list of the required parameters for calling the service:

provider

Provider name

.
field

Description of value

.
value

INE code of each province

.

Source

https://dieghernan.github.io/leaflet-providersESP/ leaflet plugin, v1.3.3.

Examples

data("leaflet.providersESP.df")

Site built with pkgdown 2.0.7.

Template rostemplate by dieghernan, based on Bootstrapious.