mapSpain package

This package provides Administrative Boundaries of Spain based on the GISCO (Geographic Information System of the Commission) Eurostat database and CartoBase SIANE from Instituto Geográfico Nacional.

Source

GISCO webpage

Details

Package	mapSpain

Type	Package

Version	0.6.0.9000

Date	2022

License	GPL-3

LazyLoad	yes

COPYRIGHT NOTICE (IGN)

This package uses data from CartoBase SIANE, provided by Instituto Geografico Nacional (IGN):

Atlas Nacional de España (ANE) CC BY 4.0 ign.es

COPYRIGHT NOTICE (GISCO)

When data downloaded from GISCO is used in any printed or electronic publication, in addition to any other provisions applicable to the whole Eurostat website, data source will have to be acknowledged in the legend of the map and in the introductory page of the publication with the following copyright notice:

EN: (C) EuroGeographics for the administrative boundaries

FR: (C) EuroGeographics pour les limites administratives

DE: (C) EuroGeographics bezüglich der Verwaltungsgrenzen

For publications in languages other than English, French or German, the translation of the copyright notice in the language of the publication shall be used.

If you intend to use the data commercially, please contact EuroGeographics for information regarding their license agreements.

References

See citation("mapSpain").

See also

Site built with pkgdown 2.0.2.

Template rostemplate by dieghernan, based on Bootstrapious.