mapSpain package

This package provides Administrative Boundaries of Spain based on the GISCO (Geographic Information System of the Commission) Eurostat database and CartoBase SIANE from Instituto Geográfico Nacional.

Source

GISCO webpage

Details

Package	mapSpain

Type	Package

Version	See sessionInfo() or DESCRIPTION file

Date	2021

License	GPL-3

LazyLoad	yes

Note

COPYRIGHT NOTICE

When data downloaded from GISCO is used in any printed or electronic publication, in addition to any other provisions applicable to the whole Eurostat website, data source will have to be acknowledged in the legend of the map and in the introductory page of the publication with the following copyright notice:

EN: (C) EuroGeographics for the administrative boundaries

FR: (C) EuroGeographics pour les limites administratives

DE: (C) EuroGeographics bezüglich der Verwaltungsgrenzen

For publications in languages other than English, French or German, the translation of the copyright notice in the language of the publication shall be used.

If you intend to use the data commercially, please contact EuroGeographics for information regarding their license agreements.

References

See citation("mapSpain").

See also

Author

dieghernan, https://github.com/dieghernan/

Site built with pkgdown 1.6.1.

Template by Bootstrapious . Ported to pkgdown by dieghernan.