mapSpain: Administrative Boundaries of Spain

Administrative Boundaries of Spain at several levels (Autonomous Communities, Provinces, Municipalities) based on the 'GISCO' 'Eurostat' database https://ec.europa.eu/eurostat/web/gisco and 'CartoBase SIANE' from 'Instituto Geografico Nacional' https://www.ign.es/. It also provides a 'leaflet' plugin and the ability of downloading and processing static tiles.

See also

Author

Maintainer: Diego Hernangómez diego.hernangomezherrero@gmail.com (ORCID) [copyright holder]

Other contributors:

Eurostat (ROR) [copyright holder]

Instituto Geográfico Nacional (ROR) [copyright holder]

Francisco J. Goerlich (ORCID) [contributor]

Site built with pkgdown 2.2.0.

Template rostemplate by dieghernan, based on Universal Theme for Hugo.